+++ mealkitary-domain/src/main/kotlin/com/mealkitary/reservation/domain/reservation/ReservationStatus.kt
@@ -1,4 +1,4 @@
-package com.mealkitary.reservation.domain
+package com.mealkitary.reservation.domain.reservation
 
 enum class ReservationStatus {
