+++ mealkitary-domain/src/test/kotlin/com/mealkitary/reservation/domain/reservation/ReservationLineItemTest.kt
@@ -1,8 +1,8 @@
-package com.mealkitary.reservation.domain
+package com.mealkitary.reservation.domain.reservation
 
 import com.mealkitary.common.model.Money
 import com.mealkitary.shop.domain.product.ProductId
-import data.ProductTestData.Companion.defaultProduct
+import data.ProductTestData
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.booleans.shouldBeTrue
@@ -20,7 +20,7 @@ class ReservationLineItemTest : AnnotationSpec() {
 
     @Test
     fun `매핑한 프로덕트와 동등성 검증`() {
-        val product = defaultProduct().build()
+        val product = ProductTestData.defaultProduct().build()
         val item = ReservationLineItem.of(ProductId(1L), "부대찌개", Money.from(1000), 2)
 
         (product == item.mapToProduct()).shouldBeTrue()
