+++ mealkitary-domain/src/test/kotlin/com/mealkitary/reservation/domain/reservation/ReservationTest.kt
@@ -1,10 +1,10 @@
-package com.mealkitary.reservation.domain
+package com.mealkitary.reservation.domain.reservation
 
 import com.mealkitary.common.model.Money
 import com.mealkitary.shop.domain.product.ProductId
 import com.mealkitary.shop.domain.shop.ShopStatus
-import data.ReservationTestData.Companion.defaultReservation
-import data.ShopTestData.Companion.defaultShop
+import data.ReservationTestData
+import data.ShopTestData
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.inspectors.forAll
@@ -23,7 +23,7 @@ class ReservationTest : AnnotationSpec() {
     @Test
     fun `예약 상품이 하나도 존재하지 않는다면 예외를 발생한다`() {
         shouldThrow<IllegalArgumentException> {
-            defaultReservation().withLineItems().build()
+            ReservationTestData.defaultReservation().withLineItems().build()
         } shouldHaveMessage "예약 상품은 적어도 한 개 이상이어야 합니다."
     }
 
@@ -37,7 +37,7 @@ class ReservationTest : AnnotationSpec() {
     @Test
     fun `예약 하려는 가게가 제공하는 예약 시간이 아니라면 예외를 발생한다`() {
         shouldThrow<IllegalArgumentException> {
-            val sut = defaultReservation().withReserveAt(
+            val sut = ReservationTestData.defaultReservation().withReserveAt(
                 LocalDateTime.of(
                     LocalDate.now().plusDays(1),
                     LocalTime.of(18, 30)
@@ -49,10 +49,10 @@ class ReservationTest : AnnotationSpec() {
 
     @Test
     fun `예약 하려는 가게가 정상 영업 중이 아니라면 예외를 발생한다`() {
-        val invalidShop = defaultShop().withStatus(ShopStatus.INVALID).build()
+        val invalidShop = ShopTestData.defaultShop().withStatus(ShopStatus.INVALID).build()
 
         shouldThrow<IllegalStateException> {
-            val sut = defaultReservation().withShop(invalidShop).build()
+            val sut = ReservationTestData.defaultReservation().withShop(invalidShop).build()
             sut.reserve()
         } shouldHaveMessage "유효하지 않은 가게입니다."
     }
@@ -133,7 +133,7 @@ class ReservationTest : AnnotationSpec() {
 
     @Test
     fun `미결제 상태가 아닌 다른 상태에서 결제를 시도하면 예외를 발생한다`() {
-        val base = defaultReservation()
+        val base = ReservationTestData.defaultReservation()
         val none = base.withReservationStatus(ReservationStatus.NONE).build()
         val paid = base.withReservationStatus(ReservationStatus.PAID).build()
         val reserved = base.withReservationStatus(ReservationStatus.RESERVED).build()
@@ -149,7 +149,7 @@ class ReservationTest : AnnotationSpec() {
 
     @Test
     fun `유효하지 않은 상품이 하나라도 존재한다면 예외를 발생한다`() {
-        val sut = defaultReservation()
+        val sut = ReservationTestData.defaultReservation()
             .withLineItems(
                 ReservationLineItem.of(ProductId(1L), "부대찌개", Money.from(1000), 2),
                 ReservationLineItem.of(ProductId(2L), "닯보끔탕", Money.from(2000), 2)
@@ -162,13 +162,13 @@ class ReservationTest : AnnotationSpec() {
 
     @Test
     fun `모든 상품과 가게, 시간이 유효하다면 예약이 가능하다`() {
-        val sut = defaultReservation().build()
+        val sut = ReservationTestData.defaultReservation().build()
         sut.reserve()
     }
 
     @Test
     fun `정상적으로 생성되지 않은 예약은 예약 거부할 수 없다`() {
-        val sut = defaultReservation().build()
+        val sut = ReservationTestData.defaultReservation().build()
 
         shouldThrow<IllegalStateException> {
             sut.reject()
@@ -177,7 +177,7 @@ class ReservationTest : AnnotationSpec() {
 
     @Test
     fun `정상적으로 생성되지 않은 예약은 수락할 수 없다`() {
-        val sut = defaultReservation().build()
+        val sut = ReservationTestData.defaultReservation().build()
 
         shouldThrow<IllegalStateException> {
             sut.accept()
@@ -196,7 +196,7 @@ class ReservationTest : AnnotationSpec() {
     @Test
     fun `예약 생성 요청을 해야 예약의 상태가 미결제로 변경된다`() {
         val sut = spyk(
-            objToCopy = defaultReservation().build(),
+            objToCopy = ReservationTestData.defaultReservation().build(),
             recordPrivateCalls = true
         )
 
@@ -207,7 +207,7 @@ class ReservationTest : AnnotationSpec() {
 
     @Test
     fun `미결제 상태인 상태가 아닌 경우 가격 계산을 할 수 없다`() {
-        val base = defaultReservation()
+        val base = ReservationTestData.defaultReservation()
         val none = base.withReservationStatus(ReservationStatus.NONE).build()
         val paid = base.withReservationStatus(ReservationStatus.PAID).build()
         val reserved = base.withReservationStatus(ReservationStatus.RESERVED).build()
@@ -223,7 +223,7 @@ class ReservationTest : AnnotationSpec() {
 
     @Test
     fun `총 상품의 가격을 계산한다`() {
-        val totalPrice = defaultReservation()
+        val totalPrice = ReservationTestData.defaultReservation()
             .withReservationStatus(ReservationStatus.NOTPAID)
             .withLineItems(
                 ReservationLineItem.of(ProductId(1L), "a", Money.from(1000), 10),
@@ -236,18 +236,18 @@ class ReservationTest : AnnotationSpec() {
     }
 
     private fun paidReservation() =
-        defaultReservation()
+        ReservationTestData.defaultReservation()
             .withReservationStatus(ReservationStatus.PAID)
             .build()
 
     private fun notPaidReservation() =
-        defaultReservation()
+        ReservationTestData.defaultReservation()
             .withReservationStatus(ReservationStatus.NOTPAID)
             .build()
 
     private fun shouldThrowWhenAcceptBeforeReservationTime(beforeTime: LocalDateTime) {
         shouldThrow<IllegalStateException> {
-            defaultReservation().withReserveAt(beforeTime).build()
+            ReservationTestData.defaultReservation().withReserveAt(beforeTime).build()
         } shouldHaveMessage "예약 시간은 적어도 미래 시점이어야 합니다."
     }
 }
