+++ mealkitary-infrastructure/adapter-persistence-spring-data-jpa/src/main/kotlin/com/mealkitary/reservation/adapter/output/persistence/ReservationRepository.kt
@@ -1,6 +1,6 @@
 package com.mealkitary.reservation.adapter.output.persistence
 
-import com.mealkitary.reservation.domain.Reservation
+import com.mealkitary.reservation.domain.reservation.Reservation
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface ReservationRepository : JpaRepository<Reservation, Long>
